+++ subjects/PE/x86/pySample/pySample_text.c
@@ -33,7 +33,7 @@ PyObject * fn100010A0(PyObject * ptrArg04, PyObject * ptrArg08)
 	int32 dwLoc04;
 	PyObject * eax_n = PyArg_ParseTuple(ptrArg08, "ii:div", &dwLoc08, &dwLoc04);
 	if (eax_n != null)
-		return Py_BuildValue("i", (int32) ((int64) dwLoc08 /32 dwLoc04));
+		return Py_BuildValue("i", (int64) dwLoc08 /32 dwLoc04);
 	return eax_n;
 }
